minigames added

+++ catpurrter/catpurrter/Shelf.cpp
@@ -24,7 +24,7 @@ void Shelf::init() {
     {400.f, 255.f}   // Position for fourth item (right, bottom shelf)
 };
 
- int idx = 0;
+ size_t idx = 0;
  for (const std::string& decoId : playerData.ownedDecorations) {
      std::string path = "assets/graphics/shelves/" + decoId + "big.png";
      sf::Texture tex;
@@ -35,7 +35,6 @@ void Shelf::init() {
              spr.setPosition(shelfPositions[idx]);
          }
          else {
-             // If too many items, stack extras at last position or offset as you wish
              spr.setPosition(shelfPositions.back());
          }
          bigDecorationSprites.push_back(spr);
